% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.comp.R
\name{surv.comp}
\alias{surv.comp}
\title{ROC-Based Methods for Comparing Survival Distributions with Right, Left, and Doubly Censored Data}
\usage{
surv.comp(
  time,
  status,
  group,
  censor_type,
  method,
  n_perm,
  progress = TRUE,
  plot = FALSE
)
}
\arguments{
\item{time}{Numeric vector of observed follow-up times (event or censoring times)
for all observations.}

\item{status}{Numeric vector indicating censoring status for each observation:
\itemize{
\item For \code{censor_type = "right"}: 0 = event, 1 = right-censored
\item For \code{censor_type = "left"}: 0 = event, -1 = left-censored
\item For \code{censor_type = "double"}: 0 = event, 1 = right-censored, -1 = left-censored.
}}

\item{group}{Numeric vector indicating the group label for each observation.
Must contain exactly two groups coded as 1 and 2.}

\item{censor_type}{Character string specifying the type of censoring.
Must be either \code{"right"}, \code{"left"}, or \code{"double"}.}

\item{method}{Character string specifying the test to perform.
Must be one of:
\itemize{
\item \code{"roc_length"}: ROC curve length-based test
\item \code{"ovl"}: overlap coefficient-based test
\item \code{"joint_method"}: joint ROC length and OVL test.
}}

\item{n_perm}{Integer specifying the number of permutation samples used to compute p-values.}

\item{progress}{Logical value indicating whether to display a progress bar
during the permutation test. Default is \code{TRUE}. If \code{FALSE},
the computation runs silently without showing progress updates.}

\item{plot}{Logical value indicating whether to generate a convex hull plot
based on the joint ROC length–OVL test. Default is \code{FALSE}.
Plotting is only available when \code{method = "joint_method"}; otherwise,
an error is returned if \code{plot = TRUE}.}
}
\value{
A list containing:
\itemize{
\item \code{message}: A character string describing the testing procedure.
\item \code{result}: A data frame with rows corresponding to the methods
(ROC Length, OVL, and/or Joint ROC Length-OVL) and columns:
\itemize{
\item \code{estimate}: Estimated values of ROC length and/or OVL
\item \code{p_value}: Permutation-based two-sided p-values.
}
\item \code{plot}: A \code{ggplot2} object showing the convex hull
visualization of the permutation distribution for the joint ROC length-OVL test.
This is returned only when \code{method = "joint_method"} and
\code{plot = TRUE}; otherwise, it is \code{NULL}.
}
}
\description{
Performs a nonparametric and semiparametric comparison of two survival distributions under
right, left, or double censoring using ROC-based metrics, including ROC curve length,
overlap coefficient (OVL), and a joint ROC length–OVL test.
}
\details{
This function implements permutation-based two-sided hypothesis tests for comparing
two survival distributions without relying on proportional hazards assumptions.
The ROC length measures global separation between survival curves, while the
overlap coefficient (OVL) quantifies distributional similarity. Density estimation is performed using nonparametric kernel methods,
where the cumulative distribution function (CDF) is estimated
using the Kaplan–Meier estimator under right censoring or the Turnbull estimator under left and double censoring.

Input validation is performed to ensure consistency of vector lengths,
censoring indicators, and method specification.
}
\examples{
\dontrun{
#### Example usage
library(ROCsurvcomp)
library(PWEXP)

# Generating right-censored data with crossing survivals
set.seed(800)
n_trt <- 100
break_trt <- c(3, 7)
rate_trt <- c(log(2)/5, log(2)/8, log(2)/13)
rate.censor_trt <- c(log(2)/20.80, log(2)/23.80, log(2)/28.80)
event_trt <- PWEXP::rpwexp(n_trt, rate = rate_trt, breakpoint = break_trt)
censor_trt <- PWEXP::rpwexp(n_trt, rate = rate.censor_trt, breakpoint = break_trt)
n_ctrl <- 100
rate_ctrl <- log(2)/8
rate.censor_ctrl <- log(2)/24
event_ctrl <- rexp(n_ctrl, rate = rate_ctrl)
censor_ctrl <- rexp(n_ctrl, rate = rate.censor_ctrl)

# Observed time and censoring status (0 = event, 1 = right-censored)
time_trt <- pmin(event_trt, censor_trt)
status_trt <- ifelse(event_trt <= censor_trt, 0, 1)
time_ctrl <- pmin(event_ctrl, censor_ctrl)
status_ctrl <- ifelse(event_ctrl <= censor_ctrl, 0, 1)
time <- c(time_trt, time_ctrl)
status <- c(status_trt, status_ctrl)
group <- c(rep(1, n_trt), rep(2, n_ctrl))

# Run `surv.comp` function
surv.comp(time = time, status = status, group = group, censor_type = "right",
          method = "roc_length", n_perm = 1000, progress = TRUE, plot = FALSE)
}
}
